Extract category page view
