+++ js/views.js
@@ -1,10 +1,9 @@
-// views.js — Navigate, dashboard, category views, and shared view composition
+// views.js — Navigate, dashboard, and shared view composition
 
 import { state } from './state.js';
-import { escapeHTML, escapeAttr, getStatus, formatDate, safeMarkerId } from './utils.js';
-import { getActiveData, filterDatesByRange, destroyAllCharts, getEffectiveRangeForDate, getLatestValueIndex, countFlagged, renderDateRangeFilter, renderChartLayersDropdown } from './data.js';
+import { escapeHTML, escapeAttr, formatDate } from './utils.js';
+import { getActiveData, destroyAllCharts, getEffectiveRangeForDate, getLatestValueIndex } from './data.js';
 import { profileStorageKey } from './profile.js';
-import { createLineChart } from './charts.js';
 import { canonicalMetric } from './wearable-adapters.js';
 import { loadContextHealthDots } from './context-cards.js';
 import { hasAIProvider, isAIPaused } from './api.js';
@@ -17,9 +16,9 @@ import { createDashboardWidgetControls } from './dashboard-widget-controls.js';
 import { createDashboardWidgetRenderers } from './dashboard-widget-renderers.js';
 import { renderFocusCard, buildFocusContext, loadFocusCard, refreshFocusCard } from './focus-card.js';
 import { configureOnboardingView, renderOnboardingBanner, renderAIConnectionReminder, dismissAIReminder, openChatProviderQuiz, setOnboardingFocus, completeOnboardingSex, completeOnboardingProfile, dismissOnboarding } from './onboarding-view.js';
-import { loadChartCardRecs } from './chart-card-recs.js';
 import { renderCategoryGlyph } from './category-glyphs.js';
 import { renderChartCard, renderTableColgroup, renderScrollableTableShell, renderTableView, renderHeatmapView, renderFattyAcidsView, renderFattyAcidsCharts } from './category-view-renderers.js';
+import { showCategory, switchView } from './category-page-view.js';
 import { configureCategoryCustomization, renameCategory, renameMarker, revertMarkerName, showEmojiPicker, changeCategoryIcon } from './category-customization.js';
 import { loadCommitHash } from './commit-hash.js';
 import { renderLightConditionsWidgetBody, renderConditionsNow, _refreshConditionsNow, _inspectConditionsNow, _setManualUvi, _clearManualUvi } from './light-conditions-now.js';
@@ -130,6 +129,8 @@ export {
   renderHeatmapView,
   renderFattyAcidsView,
   renderFattyAcidsCharts,
+  showCategory,
+  switchView,
   showLight,
   renderLightTodayStrip,
   renderLightChannelsLive,
@@ -215,32 +216,6 @@ function setupDropZone() {
 }
 
 function markerHasData(m) { return m.values?.some(v => v !== null) ?? false; }
-function sortCategoryChartEntries(entries, categoryKey) {
-  const preserved = state._preserveCategoryCardOrder;
-  if (preserved?.categoryKey === categoryKey && Array.isArray(preserved.markerKeys)) {
-    const order = new Map(preserved.markerKeys.map((key, index) => [key, index]));
-    entries.sort(([ka], [kb]) => (order.get(ka) ?? Number.MAX_SAFE_INTEGER) - (order.get(kb) ?? Number.MAX_SAFE_INTEGER));
-    delete state._preserveCategoryCardOrder;
-    return;
-  }
-  delete state._preserveCategoryCardOrder;
-
-  // Default category landing sort: markers with catalog slots first, then
-  // by status (out-of-range before normal).
-  const catalog = window._cachedCatalog;
-  const hasSlot = (k) => catalog?.slots?.[categoryKey + '.' + k] ? 0 : 1;
-  const statusOrder = { high: 0, low: 0, normal: 1, missing: 2 };
-  entries.sort(([ka, a], [kb, b]) => {
-    const slotDiff = hasSlot(ka) - hasSlot(kb);
-    if (slotDiff !== 0) return slotDiff;
-    const ai = getLatestValueIndex(a.values), bi = getLatestValueIndex(b.values);
-    const ar = ai !== -1 ? getEffectiveRangeForDate(a, ai) : { min: null, max: null };
-    const br = bi !== -1 ? getEffectiveRangeForDate(b, bi) : { min: null, max: null };
-    const as = ai !== -1 ? getStatus(a.values[ai], ar.min, ar.max) : 'missing';
-    const bs = bi !== -1 ? getStatus(b.values[bi], br.min, br.max) : 'missing';
-    return (statusOrder[as] ?? 2) - (statusOrder[bs] ?? 2);
-  });
-}
 function setDetailModalShell(...classes) {
   const modal = document.getElementById('detail-modal');
   if (!modal) return null;
@@ -765,139 +740,6 @@ export function showDashboard(data) {
   const _hasProfile = _p?.name && _p.name !== 'Default' && state.profileSex;
   if (_hasProfile && hasData) {
     if (window.startTour) window.startTour(true);
-  }
-}
-
-// ═══════════════════════════════════════════════
-// CATEGORY VIEWS
-// ═══════════════════════════════════════════════
-
-export function showCategory(categoryKey, preData) {
-  // categoryKey is interpolated into inline-onclick handlers below (rename,
-  // switchView, showDetailModal). Reject anything that doesn't
-  // match the strict allowlist so a poisoned customMarker key can't break
-  // out of the JS string context.
-  if (!safeMarkerId(categoryKey)) return;
-  // Ensure catalog is preloaded for sorting and rec links
-  if (window.loadCatalog && !window._cachedCatalog) window.loadCatalog().then(c => { window._cachedCatalog = c; });
-  const rawData = preData || getActiveData();
-  const data = filterDatesByRange(rawData);
-  const cat = data.categories[categoryKey];
-  const main = document.getElementById("main-content");
-  const allEntries = Object.entries(cat.markers).filter(([, m]) => !m.hidden);
-  const withData = allEntries.filter(([, m]) => markerHasData(m));
-  const countLabel = withData.length < allEntries.length ? `${withData.length} of ${allEntries.length} biomarkers with data` : `${allEntries.length} biomarkers tracked`;
-  const renameBtn = ` <span class="ref-edited-badge" role="button" tabindex="0" aria-label="Rename category" title="Rename category" onclick="event.stopPropagation();renameCategory('${categoryKey}')" style="cursor:pointer;font-size:12px">rename</span>`;
-  let html = `<div class="category-header"><h2>${renderCategoryGlyph(categoryKey, cat.label)}<span class="category-title-text">${escapeHTML(cat.label)}</span>${renameBtn}</h2>
-    <p>${countLabel}</p></div>`;
-
-  html += `<div style="display:flex;gap:12px;align-items:center;flex-wrap:wrap;margin-bottom:20px">`;
-  html += `<div class="view-toggle" role="tablist" aria-label="View mode" style="margin-bottom:0">
-    <button class="view-btn active" role="tab" aria-selected="true" tabindex="0" onclick="switchView('charts','${categoryKey}',this)">Charts</button>
-    <button class="view-btn" role="tab" aria-selected="false" tabindex="-1" onclick="switchView('table','${categoryKey}',this)">Table</button>
-    <button class="view-btn" role="tab" aria-selected="false" tabindex="-1" onclick="switchView('heatmap','${categoryKey}',this)">Heatmap</button></div>`;
-  html += renderDateRangeFilter();
-  html += renderChartLayersDropdown();
-  html += `</div>`;
-
-  html += `<div id="view-content">`;
-  if (withData.length === 0) {
-    html += `<div class="empty-state"><div class="empty-state-icon empty-state-icon-category">${renderCategoryGlyph(categoryKey, cat.label, { large: true })}</div>
-      <h3>No Data Available</h3><p>Import lab results containing ${escapeHTML(cat.label.toLowerCase())} markers to see data here.</p></div>`;
-  } else if (cat.singleDate) {
-    html += renderFattyAcidsView(cat, categoryKey);
-  } else {
-    sortCategoryChartEntries(withData, categoryKey);
-    html += `<div class="charts-grid">`;
-    for (const [key, marker] of withData) {
-      // Skip legacy customMarkers with unsafe keys — they can't be safely
-      // embedded in inline-onclick handlers.
-      if (!safeMarkerId(key)) continue;
-      html += renderChartCard(categoryKey + "_" + key, marker, data.dateLabels);
-    }
-    html += `</div>`;
-    // Show empty markers (no data yet) as clickable cards
-    const noData = allEntries.filter(([, m]) => !markerHasData(m));
-    if (noData.length > 0) {
-      html += `<div style="margin-top:16px"><p style="color:var(--text-secondary);font-size:13px;margin-bottom:8px">No data yet</p><div style="display:flex;flex-wrap:wrap;gap:8px">`;
-      for (const [key, marker] of noData) {
-        if (!safeMarkerId(key)) continue;
-        const id = categoryKey + '_' + key;
-        html += `<div class="chart-card" role="button" tabindex="0" aria-label="Add value for ${escapeHTML(marker.name)}" onclick="showDetailModal('${id}')" style="cursor:pointer;padding:12px 16px;min-height:auto;flex:0 0 auto">
-          <span style="color:var(--text-secondary)">${escapeHTML(marker.name)}</span>
-          <span style="color:var(--text-muted);font-size:11px;margin-left:6px">+ add value</span></div>`;
-      }
-      html += `</div></div>`;
-    }
-  }
-  html += `</div>`;
-  main.innerHTML = html;
-
-  const savedView = state.categoryView;
-  if (savedView === 'table' || savedView === 'heatmap') {
-    const buttons = main.querySelectorAll('.view-toggle .view-btn');
-    const idx = savedView === 'table' ? 1 : 2;
-    if (buttons[idx]) { switchView(savedView, categoryKey, buttons[idx]); return; }
-  }
-
-  if (withData.length === 0) { /* no charts to render */ }
-  else if (cat.singleDate) { renderFattyAcidsCharts(cat); }
-  else {
-    for (const [key, marker] of withData) {
-      createLineChart(categoryKey + "_" + key, marker, data.dateLabels, data.dates, data.phaseLabels);
-    }
-  }
-  loadChartCardRecs();
-}
-
-export function switchView(view, categoryKey, btn) {
-  // categoryKey reaches inline-onclick handlers via renderChartCard /
-  // renderFattyAcidsView / renderTableView / renderHeatmapView. Same
-  // allowlist guard as showCategory.
-  if (!safeMarkerId(categoryKey)) return;
-  state.categoryView = view;
-  document.querySelectorAll(".view-btn").forEach(b => {
-    b.classList.remove("active");
-    b.setAttribute('aria-selected', 'false');
-    b.setAttribute('tabindex', '-1');
-  });
-  btn.classList.add("active");
-  btn.setAttribute('aria-selected', 'true');
-  btn.setAttribute('tabindex', '0');
-  destroyAllCharts();
-  const rawData = getActiveData();
-  const data = filterDatesByRange(rawData);
-  const cat = data.categories[categoryKey];
-  const container = document.getElementById("view-content");
-  // Pre-sanitize date labels at the call boundary — CodeQL's taint analysis
-  // (js/xss-through-dom) doesn't trace sanitizers across function calls, so
-  // even though renderTableView/renderHeatmapView re-escape internally,
-  // escaping here closes the call-site taint flow. Date arrays stay raw
-  // because they're consumed by JSON.stringify in inline-onclick attrs
-  // (escapeHTML would double-escape the JSON literal).
-  const safeLabels = Array.isArray(data.dateLabels) ? data.dateLabels.map(escapeHTML) : data.dateLabels;
-  if (view === "table") {
-    container.innerHTML = renderTableView(cat, safeLabels, categoryKey, data.dates);
-  } else if (view === "heatmap") {
-    container.innerHTML = renderHeatmapView(cat, safeLabels, data.dates, categoryKey);
-  } else {
-    if (cat.singleDate) {
-      container.innerHTML = renderFattyAcidsView(cat, categoryKey);
-      renderFattyAcidsCharts(cat);
-    } else {
-      // Per-key safety check skips legacy customMarkers with unsafe keys so
-      // they never reach inline-onclick handlers in renderChartCard.
-      const withData = Object.entries(cat.markers).filter(([key, m]) => markerHasData(m) && safeMarkerId(key));
-      let html = `<div class="charts-grid">`;
-      for (const [key, marker] of withData) {
-        html += renderChartCard(categoryKey + "_" + key, marker, data.dateLabels);
-      }
-      html += `</div>`;
-      container.innerHTML = html;
-      for (const [key, marker] of withData) {
-        createLineChart(categoryKey + "_" + key, marker, data.dateLabels, data.dates, data.phaseLabels);
-      }
-    }
   }
 }
