+++ tests/test-correctness-phase2.js
@@ -84,6 +84,7 @@ const pwaAppShellAssets = [
   '/js/dashboard-widget-renderers.js',
   '/js/chart-card-recs.js',
   '/js/category-glyphs.js',
+  '/js/category-page-view.js',
   '/js/category-view-renderers.js',
   '/js/category-customization.js',
   '/js/commit-hash.js',
